requirement.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
adafruit-circuitpython-neopixel==6.3.9
RPi.GPIO==0.7.1
gpiozero==1.6.2
